lifeos/requirements.txt
Flask>=3.0,<4
Flask-SQLAlchemy>=3.1
Flask-Migrate>=4.0
Flask-JWT-Extended>=4.6
Flask-Limiter>=3.5
Flask-Bcrypt>=1.0
Flask-Login>=0.6
Alembic>=1.13
gunicorn>=21.2
pydantic>=2.5
pytest>=8.2
rich>=13.7
email-validator>=2.1
joblib>=1.3
python-dotenv>=1.0
redis>=4.3
psycopg2-binary>=2.9
